Renamed CertService to -handler. Rebuilt to proper DI
